aider: Add ability to run Dockerfile by updating docker-run.sh script.
# Aider chat conversation:

USER: I added these *read-write* files: package-lock.json, package.json

ASSISTANT: It seems you have added the `package-lock.json` and `package.json` files to the list of *read-write* files. However, you haven't specified any changes to be made to these files. If you need any modifications or updates to these files, please provide the details of the changes you would like to apply.

USER: add the ability to run d`Dockerfile`

ASSISTANT: To add the ability to run the Dockerfile, we need to update the `scripts/docker-run.sh` script to build the Docker image without specifying a target stage (since the Dockerfile does not define multiple stages) and to run the Docker container. Here are the changes:

```diff
--- scripts/docker-run.sh
+++ scripts/docker-run.sh
@@ ... @@
 set -e
-docker build . --target processor -t squid-processor
+docker build . -t squid-processor
 # make sure the port matches .env.
 # For Linux, add --add-host=host.docker.internal:host-gateway
-docker run --rm -e DB_HOST=host.docker.internal --env-file=.env squid-processor
+docker run --rm -e DB_HOST=host.docker.internal --env-file=.env -p 4350:4350 squid-processor
```

These changes will build the Docker image using the Dockerfile and run the container, mapping the port 4350 from the container to the host, which is the port the application listens on as specified in the Dockerfile.