aider: Added ability to run Docker from `package.json`.
# Aider chat conversation:

USER: I added these *read-write* files: .env

ASSISTANT: It appears that you have indicated the `.env` file as a *read-write* file, but you have not provided any specific instructions or requests for changes to be made to this file. If you require any updates or modifications to the `.env` file, please provide the details of the changes you would like to implement.

USER: add the ability to run docker from `package.json`

ASSISTANT: To add the ability to run Docker from the `package.json` file, we can add scripts to build and run the Docker container. Here are the changes to the `package.json` file:

```diff
--- package.json
+++ package.json
@@ ... @@
   "scripts": {
     "build": "rm -rf lib && tsc",
+    "docker:build": "docker build . -t squid-processor",
+    "docker:run": "docker run --rm -e DB_HOST=host.docker.internal --env-file=.env -p 4350:4350 squid-processor",
     "ctag": "ctags package.json src/* README.md schema.graphql squid.yaml scripts/* commands.json .env run.sh tsconfig.json typegen.json docker-compose.yml "
   },
   "dependencies": {
```

With these changes, you can now use `npm run docker:build` to build the Docker image and `npm run docker:run` to run the Docker container.